+++ test/function/algebra/simplify.test.js
@@ -43,6 +43,38 @@ describe('simplify', function() {
     simplifyAndCompare('(-1)*x', '-x');
   });
 
+  it('should handle function assignments', function() {
+    var node = math.expression.node;
+    var f = new node.FunctionAssignmentNode('sigma', ['x'], math.parse('1 / (1 + exp(-x))'));
+    assert.equal(f.toString(), 'sigma(x) = 1 / (1 + exp(-x))');
+    assert.equal(f.eval()(5), 0.9933071490757153);
+    var fsimplified = math.simplify.simplifyCore(f);
+    assert.equal(fsimplified.toString(), 'sigma(x) = 1 / (1 + exp(-x))');
+    assert.equal(fsimplified.eval()(5), 0.9933071490757153);
+  });
+
+  it('should handle custom functions', function() {
+    function doubleIt (x) { return x + x }
+    var node = math.expression.node;
+    var f = new node.FunctionNode(new node.SymbolNode('doubleIt'), [new node.SymbolNode('value')]);
+    assert.equal(f.toString(), 'doubleIt(value)');
+    assert.equal(f.eval({ doubleIt: doubleIt, value: 4 }), 8);
+    var fsimplified = math.simplify.simplifyCore(f);
+    assert.equal(fsimplified.toString(), 'doubleIt(value)');
+    assert.equal(fsimplified.eval({ doubleIt: doubleIt, value: 4 }), 8);
+  });
+
+  it('should handle immediately invoked function assignments', function() {
+    var node = math.expression.node;
+    var s = new node.FunctionAssignmentNode('sigma', ['x'], math.parse('1 / (1 + exp(-x))'));
+    var f = new node.FunctionNode(s, [new node.SymbolNode('x')]);
+    assert.equal(f.toString(), '(sigma(x) = 1 / (1 + exp(-x)))(x)');
+    assert.equal(f.eval({x: 5}), 0.9933071490757153);
+    var fsimplified = math.simplify.simplifyCore(f);
+    assert.equal(fsimplified.toString(), '(sigma(x) = 1 / (1 + exp(-x)))(x)');
+    assert.equal(fsimplified.eval({x: 5}), 0.9933071490757153);
+  })
+
   it('should simplify (n- -n1)', function() {
     simplifyAndCompare('2 + -3', '-1');
     simplifyAndCompare('2 - 3', '-1');