+++ test/type/unit/Unit.test.js
@@ -252,6 +252,12 @@ describe('Unit', function() {
       var u = new Unit(math.fraction(1,3), 'cm');
       assert.deepEqual(u.toNumeric('mm'), math.fraction(10,3));
     });
+    it ('should simplify units before returning a numeric value', function () {
+      var cm = new Unit(1, 'gram');
+      var m = new Unit(1, 'kilogram');
+      var product = cm.multiply(m)
+      assert.deepEqual(product.toNumeric(), 0.001);
+    });
   });
 
   describe('to', function() {